requirements.txt
torch
torchvision
transformers>=4.57.0
accelerate
peft
bitsandbytes
datasets
pillow
opencv-python
numpy
pandas
pyyaml
spacy
easyocr
scikit-learn
tqdm
pydantic
sentencepiece
safetensors
